chore(restructuring): adds the model, repo, service, and controller packages and moved ItemController into model package
