+++ src/main/java/com/mrxgrc/inventory/controller/ItemController.java
@@ -1,4 +1,4 @@
-package com.mrxgrc.inventory;
+package com.mrxgrc.inventory.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
