+++ src/test/java/com/mrxgrc/inventory/ItemControllerTest.java
@@ -1,5 +1,6 @@
 package com.mrxgrc.inventory;
 
+import com.mrxgrc.inventory.controller.ItemController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
